+++ app/server.ts
@@ -41,6 +41,7 @@ import communityReport from './backend/lib/communityReportsDueDate';
 import map from './backend/lib/map/map';
 import dashboardExport from './backend/lib/dashboard/dashboard_export';
 import intake from './backend/lib/intake';
+import cbc from './backend/lib/cbc/cbc';
 
 // Function to exclude middleware from certain routes
 // The paths argument takes an array of strings containing routes to exclude from the middleware
@@ -161,6 +162,7 @@ app.prepare().then(async () => {
   server.use('/', map);
   server.use('/', dashboardExport);
   server.use('/', intake);
+  server.use('/', cbc);
 
   server.all('*', async (req, res) => handle(req, res));